taebbong_drf/06_MyBoard/requirements.txt
asgiref==3.5.2
Django==4.1.2
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
